Move controllers to corresponding packages
