+++ spring-boot-admin-server/src/main/java/de/codecentric/boot/admin/journal/web/JournalController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.codecentric.boot.admin.controller;
+package de.codecentric.boot.admin.journal.web;
 
 import static java.util.Collections.synchronizedCollection;
