+++ spring-boot-admin-server/src/main/java/de/codecentric/boot/admin/registry/web/RegistryController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.codecentric.boot.admin.controller;
+package de.codecentric.boot.admin.registry.web;
 
 import java.util.Collection;
